+++ designer/core/__init__.py
@@ -1,4 +1,4 @@
 from designer.core.director import *
 from designer.core.event import *
 from designer.core.window import Window
-from designer.core.internal_image import InternalImage
+from designer.core.internal_image import InternalImage, DesignerSurface